auth_service/src/requirements.txt
uvicorn==0.12.2
uvloop==0.14.0
flake8==4.0
python-dotenv==0.20.0
psycopg2-binary==2.9.3
gunicorn==20.1.0
Flask==2.0.2
gevent==21.8.0
redis==4.1.0
Flask-JWT-Extended==4.3.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
pydantic==1.9.0
flask_apispec==0.11.1
apispec_webframeworks==0.5.2
pyyaml==6.0
flask_swagger_ui==4.11.1
flask_migrate==3.1.0
opentelemetry-api==1.10.0
opentelemetry-sdk==1.10.0
opentelemetry-instrumentation-flask==0.29b1
opentelemetry-exporter-jaeger==1.10.0
authlib==1.0.1
requests==2.27.1